update prompts and order
