+++ challenges/aot/2023/2/prompt.md
@@ -1,15 +1,9 @@
-### The Gift Wrapper
+## The Gift Wrapper
 
-Santa's workshop needs an updated type for gift wrappers! Currently, there's a basic type defined as follows:
+Did you know that there's also monetary inflation at the North Pole? You betcha, there is. And after 200+ years without a pay raise, the Santa's elves are beginning to discuss a general strike.
 
-```typescript
-type GiftWrapper = {
-  present: any;
-  from: any;
-  to: any;
-};
-```
+December 2nd is just about the worst time imaginable for such a strike, and Santa's desparate to calm the elves down. If he can just wrap a few presents maybe the elves will forget that they're being paid well below market rate (don't worry: the North Pole's actually still a Deleware-based startup so therefore it's ok).
 
-Your task is to create a type named `GiftWrapper` that takes in three type parameters: `Present`, `From`, and `To`. The goal is to modify the existing `GiftWrapper` type to enforce specific types for each property based on the provided type parameters.
+There's a `GiftWrapper` type to help keep the wrapping process organized, but it needs something... it needs some way to be paramaterized. What we have so far is nice as a _generic_ (\*wink wink\*) starting point... but it needs some way to provide specific values for `Present`, `From`, and `To` at the type layer..
 
-Good luck!
+Please help! Otherwise the reindeer might catch wind of this and start a striek of their own in solidarity with the elves!
