+++ challenges/aot/2023/22/user.ts
@@ -1,17 +1,30 @@
-type Connect4Cell = "🔴" | "🟡" | "  ";
+/** because "dashing" implies speed */
+type Dasher = '💨';
 
-type EmptyBoard = [
-  ["  ", "  ", "  ", "  ", "  ", "  ", "  "],
-  ["  ", "  ", "  ", "  ", "  ", "  ", "  "],
-  ["  ", "  ", "  ", "  ", "  ", "  ", "  "],
-  ["  ", "  ", "  ", "  ", "  ", "  ", "  "],
-  ["  ", "  ", "  ", "  ", "  ", "  ", "  "],
-  ["  ", "  ", "  ", "  ", "  ", "  ", "  "],
-];
+/** representing dancing or grace */
+type Dancer = '💃';
 
-type NewGame = {
-  board: EmptyBoard;
-  state: "🟡";
-};
+/** a deer, prancing */
+type Prancer = '🦌';
 
-type Connect4 = unknown;
+/** a star for the dazzling, slightly mischievous Vixen */
+type Vixen = '🌟';
+
+/** for the celestial body that shares its name */
+type Comet = '☄️';
+
+/** symbolizing love, as Cupid is the god of love */
+type Cupid = '❤️';
+
+/** representing thunder, as "Donner" means thunder in German */
+type Donner = '🌩️';
+
+/** meaning lightning in German, hence the lightning bolt */
+type Blitzen = '⚡';
+
+/** for his famous red nose */
+type Rudolph = '🔴';
+
+type Reindeer = Dasher | Dancer | Prancer | Vixen | Comet | Cupid | Donner | Blitzen | Rudolph;
+
+type Validate = unknown;
