+++ challenges/aot/2023/23/tests.ts
@@ -1,85 +1,211 @@
-import { Equal, Expect } from 'type-testing';
+import { Expect, Equal } from "type-testing";
 
-type test_sudoku_1_actual = Validate<[
+type test_move1_actual = Connect4<NewGame, 0>;
 //   ^?
-  [['💨', '💃', '🦌'], ['☄️', '❤️', '🌩️'], ['🌟', '⚡', '🔴']],
-  [['🌟', '⚡', '🔴'], ['💨', '💃', '🦌'], ['☄️', '❤️', '🌩️']],
-  [['☄️', '❤️', '🌩️'], ['🌟', '⚡', '🔴'], ['💨', '💃', '🦌']],
-  [['🦌', '💨', '💃'], ['⚡', '☄️', '❤️'], ['🔴', '🌩️', '🌟']],
-  [['🌩️', '🔴', '🌟'], ['🦌', '💨', '💃'], ['⚡', '☄️', '❤️']],
-  [['⚡', '☄️', '❤️'], ['🌩️', '🔴', '🌟'], ['🦌', '💨', '💃']],
-  [['💃', '🦌', '💨'], ['❤️', '🌟', '☄️'], ['🌩️', '🔴', '⚡']],
-  [['🔴', '🌩️', '⚡'], ['💃', '🦌', '💨'], ['❤️', '🌟', '☄️']],
-  [['❤️', '🌟', '☄️'], ['🔴', '🌩️', '⚡'], ['💃', '🦌', '💨']]
-]>;
-type test_sudoku_1 = Expect<Equal<test_sudoku_1_actual, true>>;
+type test_move1_expected = {
+  board: [
+    ["  ", "  ", "  ", "  ", "  ", "  ", "  "],
+    ["  ", "  ", "  ", "  ", "  ", "  ", "  "],
+    ["  ", "  ", "  ", "  ", "  ", "  ", "  "],
+    ["  ", "  ", "  ", "  ", "  ", "  ", "  "],
+    ["  ", "  ", "  ", "  ", "  ", "  ", "  "],
+    ["🟡", "  ", "  ", "  ", "  ", "  ", "  "],
+  ];
+  state: "🔴";
+};
+type test_move1 = Expect<Equal<test_move1_actual, test_move1_expected>>;
 
-type test_sudoku_2_actual = Validate<[
+type test_move2_actual = Connect4<test_move1_actual, 0>;
 //   ^?
-  [['🌩️', '💨', '☄️'], ['🌟', '🦌', '⚡'], ['❤️', '🔴', '💃']],
-  [['🌟', '⚡', '❤️'], ['🔴', '💃', '☄️'], ['🌩️', '💨', '🦌']],
-  [['🔴', '🦌', '💃'], ['💨', '❤️', '🌩️'], ['🌟', '⚡', '☄️']],
-  [['❤️', '☄️', '🌩️'], ['💃', '⚡', '🔴'], ['💨', '🦌', '🌟']],
-  [['🦌', '💃', '⚡'], ['🌩️', '🌟', '💨'], ['🔴', '☄️', '❤️']],
-  [['💨', '🌟', '🔴'], ['🦌', '☄️', '❤️'], ['⚡', '💃', '🌩️']],
-  [['☄️', '🔴', '💨'], ['❤️', '🌩️', '🦌'], ['💃', '🌟', '⚡']],
-  [['💃', '❤️', '🦌'], ['⚡', '🔴', '🌟'], ['☄️', '🌩️', '💨']],
-  [['⚡', '🌩️', '🌟'], ['☄️', '💨', '💃'], ['🦌', '❤️', '🔴']]
-]>;
-type test_sudoku_2 = Expect<Equal<test_sudoku_2_actual, true>>;
+type test_move2_expected = {
+  board: [
+    ["  ", "  ", "  ", "  ", "  ", "  ", "  "],
+    ["  ", "  ", "  ", "  ", "  ", "  ", "  "],
+    ["  ", "  ", "  ", "  ", "  ", "  ", "  "],
+    ["  ", "  ", "  ", "  ", "  ", "  ", "  "],
+    ["🔴", "  ", "  ", "  ", "  ", "  ", "  "],
+    ["🟡", "  ", "  ", "  ", "  ", "  ", "  "],
+  ];
+  state: "🟡";
+};
+type test_move2 = Expect<Equal<test_move2_actual, test_move2_expected>>;
 
-type test_sudoku_3_actual = Validate<[
+type test_move3_actual = Connect4<test_move2_actual, 0>;
 //   ^?
-  [['🦌', '🔴', '💃'], ['🌩️', '☄️', '💨'], ['⚡', '❤️', '🌟']],
-  [['🌟', '⚡', '💨'], ['❤️', '💃', '🔴'], ['☄️', '🌩️', '🦌']],
-  [['☄️', '🌩️', '❤️'], ['⚡', '🌟', '🦌'], ['💃', '🔴', '💨']],
-  [['🌩️', '💃', '🔴'], ['🦌', '💨', '⚡'], ['🌟', '☄️', '❤️']],
-  [['❤️', '☄️', '⚡'], ['💃', '🌩️', '🌟'], ['🦌', '💨', '🔴']],
-  [['💨', '🌟', '🦌'], ['☄️', '🔴', '❤️'], ['🌩️', '💃', '⚡']],
-  [['💃', '💨', '🌟'], ['🔴', '🦌', '☄️'], ['❤️', '⚡', '🌩️']],
-  [['🔴', '❤️', '☄️'], ['🌟', '⚡', '🌩️'], ['💨', '🦌', '💃']],
-  [['⚡', '🦌', '🌩️'], ['💨', '❤️', '💃'], ['🔴', '🌟', '☄️']]
-]>;
-type test_sudoku_3 = Expect<Equal<test_sudoku_3_actual, true>>;
+type test_move3_expected = {
+  board: [
+    ["  ", "  ", "  ", "  ", "  ", "  ", "  "],
+    ["  ", "  ", "  ", "  ", "  ", "  ", "  "],
+    ["  ", "  ", "  ", "  ", "  ", "  ", "  "],
+    ["🟡", "  ", "  ", "  ", "  ", "  ", "  "],
+    ["🔴", "  ", "  ", "  ", "  ", "  ", "  "],
+    ["🟡", "  ", "  ", "  ", "  ", "  ", "  "],
+  ];
+  state: "🔴";
+};
+type test_move3 = Expect<Equal<test_move3_actual, test_move3_expected>>;
 
-type test_sudoku_4_actual = Validate<[
+type test_move4_actual = Connect4<test_move3_actual, 1>;
 //   ^?
-  [['💨', '💃', '🦌'], ['☄️', '❤️', '🌩️'], ['🌟', '⚡', '🔴']],
-  [['🌟', '⚡', '🔴'], ['💨', '💃', '🦌'], ['☄️', '❤️', '🌩️']],
-  [['☄️', '❤️', '🌩️'], ['🌟', '⚡', '🔴'], ['💨', '💃', '🦌']],
-  [['🦌', '💨', '💃'], ['⚡', '☄️', '❤️'], ['🔴', '🌩️', '🌟']],
-  [['🌩️', '🔴', '🌟'], ['🦌', '💨', '💃'], ['⚡', '☄️', '❤️']],
-  [['⚡', '☄️', '❤️'], ['🌩️', '🔴', '🌟'], ['🦌', '💨', '💃']],
-  [['💃', '🦌', '💨'], ['❤️', '🌟', '☄️'], ['⚡', '🔴', '🌟']],
-  [['🔴', '🌩️', '⚡'], ['💃', '🦌', '💨'], ['❤️', '🌟', '☄️']],
-  [['❤️', '🌟', '☄️'], ['🔴', '🌩️', '⚡'], ['💃', '🦌', '💨']]
-]>;
-type test_sudoku_4 = Expect<Equal<test_sudoku_4_actual, false>>;
+type test_move4_expected = {
+  board: [
+    ["  ", "  ", "  ", "  ", "  ", "  ", "  "],
+    ["  ", "  ", "  ", "  ", "  ", "  ", "  "],
+    ["  ", "  ", "  ", "  ", "  ", "  ", "  "],
+    ["🟡", "  ", "  ", "  ", "  ", "  ", "  "],
+    ["🔴", "  ", "  ", "  ", "  ", "  ", "  "],
+    ["🟡", "🔴", "  ", "  ", "  ", "  ", "  "],
+  ];
+  state: "🟡";
+};
+type test_move4 = Expect<Equal<test_move4_actual, test_move4_expected>>;
 
-type test_sudoku_5_actual = Validate<[
+type test_move5_actual = Connect4<test_move4_actual, 2>;
 //   ^?
-  [['🌩️', '💨', '☄️'], ['🌟', '🦌', '⚡'], ['❤️', '🔴', '💃']],
-  [['🌟', '⚡', '❤️'], ['🔴', '💃', '☄️'], ['🌩️', '💨', '🦌']],
-  [['🔴', '🦌', '💃'], ['💨', '❤️', '🌩️'], ['🌟', '⚡', '☄️']],
-  [['❤️', '☄️', '🌩️'], ['💃', '⚡', '🔴'], ['💨', '🦌', '🌟']],
-  [['🦌', '💃', '⚡'], ['🌩️', '🌟', '💨'], ['🔴', '☄️', '❤️']],
-  [['💨', '🌟', '🔴'], ['🦌', '☄️', '❤️'], ['⚡', '💃', '🌩️']],
-  [['☄️', '🔴', '💨'], ['❤️', '💃', '🦌'], ['💃', '🌟', '⚡']],
-  [['💃', '❤️', '🦌'], ['⚡', '🔴', '🌟'], ['☄️', '🌩️', '💨']],
-  [['⚡', '🌩️', '🌟'], ['☄️', '💨', '💃'], ['🦌', '❤️', '🔴']]
-]>;
-type test_sudoku_5 = Expect<Equal<test_sudoku_5_actual, false>>;
+type test_move5_expected = {
+  board: [
+    ["  ", "  ", "  ", "  ", "  ", "  ", "  "],
+    ["  ", "  ", "  ", "  ", "  ", "  ", "  "],
+    ["  ", "  ", "  ", "  ", "  ", "  ", "  "],
+    ["🟡", "  ", "  ", "  ", "  ", "  ", "  "],
+    ["🔴", "  ", "  ", "  ", "  ", "  ", "  "],
+    ["🟡", "🔴", "🟡", "  ", "  ", "  ", "  "],
+  ];
+  state: "🔴";
+};
+type test_move5 = Expect<Equal<test_move5_actual, test_move5_expected>>;
 
-type test_sudoku_6_actual = Validate<[
+type test_move6_actual = Connect4<test_move5_actual, 1>;
 //   ^?
-  [['⚡', '🔴', '🌩️'], ['🦌', '❤️', '💨'], ['💨', '🌟', '☄️']],
-  [['❤️', '🦌', '🌟'], ['💨', '🌟', '🔴'], ['💃', '⚡', '🌩️']],
-  [['💨', '💃', '🌟'], ['☄️', '⚡', '🌩️'], ['🔴', '❤️', '🦌']],
-  [['🦌', '⚡', '🔴'], ['❤️', '💃', '💨'], ['☄️', '🌩️', '🌟']],
-  [['🌟', '🌩️', '💃'], ['⚡', '🔴', '☄️'], ['❤️', '🦌', '💨']],
-  [['☄️', '💨', '❤️'], ['🌟', '🌩️', '🦌'], ['⚡', '💃', '🔴']],
-  [['🌩️', '☄️', '💨'], ['💃', '🦌', '⚡'], ['🌟', '🔴', '❤️']],
-  [['🔴', '❤️', '⚡'], ['🌩️', '☄️', '🌟'], ['🦌', '💨', '💃']],
-  [['💃', '🌟', '🦌'], ['🔴', '💨', '❤️'], ['🌩️', '☄️', '⚡']]
-]>;
-type test_sudoku_6 = Expect<Equal<test_sudoku_6_actual, false>>;
+type test_move6_expected = {
+  board: [
+    ["  ", "  ", "  ", "  ", "  ", "  ", "  "],
+    ["  ", "  ", "  ", "  ", "  ", "  ", "  "],
+    ["  ", "  ", "  ", "  ", "  ", "  ", "  "],
+    ["🟡", "  ", "  ", "  ", "  ", "  ", "  "],
+    ["🔴", "🔴", "  ", "  ", "  ", "  ", "  "],
+    ["🟡", "🔴", "🟡", "  ", "  ", "  ", "  "],
+  ];
+  state: "🟡";
+};
+type test_move6 = Expect<Equal<test_move6_actual, test_move6_expected>>;
+
+type test_red_win_actual = Connect4<
+  {
+    board: [
+      ['  ', '  ', '  ', '  ', '  ', '  ', '  '],
+      ['  ', '  ', '  ', '  ', '  ', '  ', '  '],
+      ['🟡', '  ', '  ', '  ', '  ', '  ', '  '],
+      ['🟡', '  ', '  ', '  ', '  ', '  ', '  '],
+      ['🔴', '🔴', '🔴', '  ', '  ', '  ', '  '],
+      ['🟡', '🔴', '🟡', '🟡', '  ', '  ', '  ']
+    ];
+    state: '🔴';
+  },
+  3
+>;
+
+type test_red_win_expected = {
+  board: [
+    ['  ', '  ', '  ', '  ', '  ', '  ', '  '],
+    ['  ', '  ', '  ', '  ', '  ', '  ', '  '],
+    ['🟡', '  ', '  ', '  ', '  ', '  ', '  '],
+    ['🟡', '  ', '  ', '  ', '  ', '  ', '  '],
+    ['🔴', '🔴', '🔴', '🔴', '  ', '  ', '  '],
+    ['🟡', '🔴', '🟡', '🟡', '  ', '  ', '  ']
+  ];
+  state: '🔴 Won';
+};
+
+type test_red_win = Expect<Equal<test_red_win_actual, test_red_win_expected>>;
+
+type test_yellow_win_actual = Connect4<
+  {
+    board: [
+      ['  ', '  ', '  ', '  ', '  ', '  ', '  '],
+      ['  ', '  ', '  ', '  ', '  ', '  ', '  '],
+      ['🔴', '  ', '  ', '  ', '  ', '  ', '  '],
+      ['🟡', '  ', '  ', '  ', '  ', '  ', '  '],
+      ['🔴', '  ', '🔴', '🔴', '  ', '  ', '  '],
+      ['🟡', '  ', '🟡', '🟡', '  ', '  ', '  ']
+    ];
+    state: '🟡';
+  },
+  1
+>;
+
+type test_yellow_win_expected = {
+  board: [
+    ['  ', '  ', '  ', '  ', '  ', '  ', '  '],
+    ['  ', '  ', '  ', '  ', '  ', '  ', '  '],
+    ['🔴', '  ', '  ', '  ', '  ', '  ', '  '],
+    ['🟡', '  ', '  ', '  ', '  ', '  ', '  '],
+    ['🔴', '  ', '🔴', '🔴', '  ', '  ', '  '],
+    ['🟡', '🟡', '🟡', '🟡', '  ', '  ', '  ']
+  ];
+  state: '🟡 Won';
+};
+
+type test_yellow_win = Expect<
+  Equal<test_yellow_win_actual, test_yellow_win_expected>
+>;
+
+type test_diagonal_yellow_win_actual = Connect4<
+  {
+    board: [
+      ['  ', '  ', '  ', '  ', '  ', '  ', '  '],
+      ['  ', '  ', '  ', '  ', '  ', '  ', '  '],
+      ['  ', '  ', '  ', '  ', '  ', '  ', '  '],
+      ['  ', '  ', '🟡', '🔴', '  ', '  ', '  '],
+      ['🔴', '🟡', '🔴', '🔴', '  ', '  ', '  '],
+      ['🟡', '🔴', '🟡', '🟡', '  ', '  ', '  ']
+    ];
+    state: '🟡';
+  },
+  3
+>;
+
+type test_diagonal_yellow_win_expected = {
+  board: [
+    ['  ', '  ', '  ', '  ', '  ', '  ', '  '],
+    ['  ', '  ', '  ', '  ', '  ', '  ', '  '],
+    ['  ', '  ', '  ', '🟡', '  ', '  ', '  '],
+    ['  ', '  ', '🟡', '🔴', '  ', '  ', '  '],
+    ['🔴', '🟡', '🔴', '🔴', '  ', '  ', '  '],
+    ['🟡', '🔴', '🟡', '🟡', '  ', '  ', '  ']
+  ];
+  state: '🟡 Won';
+};
+
+type test_diagonal_yellow_win = Expect<
+  Equal<test_diagonal_yellow_win_actual, test_diagonal_yellow_win_expected>
+>;
+
+type test_draw_actual = Connect4<
+  {
+    board: [
+      ['🟡', '🔴', '🔴', '🟡', '🟡', '🔴', '  '],
+      ['🔴', '🟡', '🟡', '🔴', '🔴', '🟡', '🔴'],
+      ['🟡', '🔴', '🔴', '🟡', '🟡', '🔴', '🟡'],
+      ['🔴', '🟡', '🟡', '🔴', '🔴', '🟡', '🔴'],
+      ['🟡', '🔴', '🔴', '🟡', '🟡', '🔴', '🟡'],
+      ['🔴', '🟡', '🟡', '🔴', '🔴', '🟡', '🔴']
+    ];
+    state: '🟡';
+  },
+  6
+>;
+
+type test_draw_expected = {
+  board: [
+    ['🟡', '🔴', '🔴', '🟡', '🟡', '🔴', '🟡'],
+    ['🔴', '🟡', '🟡', '🔴', '🔴', '🟡', '🔴'],
+    ['🟡', '🔴', '🔴', '🟡', '🟡', '🔴', '🟡'],
+    ['🔴', '🟡', '🟡', '🔴', '🔴', '🟡', '🔴'],
+    ['🟡', '🔴', '🔴', '🟡', '🟡', '🔴', '🟡'],
+    ['🔴', '🟡', '🟡', '🔴', '🔴', '🟡', '🔴']
+  ];
+  state: 'Draw';
+};
+
+type test_draw = Expect<Equal<test_draw_actual, test_draw_expected>>;
