+++ challenges/aot/2023/5/prompt.md
@@ -1,29 +1,25 @@
-### Organize Santa's List
+## Organize Santa's List
 
-Santa's elves are working on organizing Santa's list, and they need your help in creating a TypeScript type to merge and sort the good and bad lists for the year. Currently, there's a type defined with placeholders:
+_\[The elves walk into work early on the morning of December 5th. A sign that reads "we're all about passion, not just paychecks" hangs above the entrance to the factory floor.\]_
 
-```typescript
-type SantasList<GoodList, BadList> = any;
+It's been a tough year for Santa's workshop. The elves are a little behind schedule on getting Santa his list. Santa _reallly really_ likes to see the full list of names far in advance of Christmas Eve when he makes his deliveries.
+
+Normally the elves get lists like this
+
+```ts
+const badList = ["Tommy", "Trash", "Queen Blattaria", /* ... many more ... */];
+const goodList = ["Jon", "David", "Captain Spectacular", /* ... many more ... */];
 ```
 
-Your challenge is to modify the `SantasList` type so that it concatenates the `GoodList` and `BadList` types.
+And they copy-pasta all the values into a TypeScript type to provide to Santa like this
 
-Here's a test case that your `SantasList` type must pass:
-
-```typescript
-// Your task: Define the SantasList type here to pass the test cases
-
-// Test cases
-const bads = ['tommy', 'trash'] as const;
-const goods = ['bash', 'tru'] as const;
-
-type cases = [
-  // Test cases here
+```ts
+type SantasList = [
+  "Tommy", "Trash", "Queen Blattaria", /* ... many more ... */
+  "Jon", "David", "Captain Spectacular", /* ... many more ... */
 ];
-
-// This should throw a TypeScript error
-// @ts-expect-error
-type error = SantasList<null, undefined>;
 ```
 
-Good luck!
+But there's a problem.. There's one elf on the team, Frymagen, that constantly reminds the others how incredible his Vim skills are. So he has always done it in years past. However this year, Frymagen got one of those MacBook Pros without the escape key and his Vim speed is drastically reduced. We need to find a better way to get Santa his list.
+
+Let's implement `SantasList` such that it can be passed the types for the `badList` and `goodList` and it will return a TypeScript tuple with the values of both lists combined.
