+++ challenges/aot/2023/8/prompt.md
@@ -1,11 +1,31 @@
-### Santa's Naughty List Filter Puzzle
+## Filtering The Children (part 3)
 
-Santa's workshop needs a type that filters out misbehaving people from his list. You're provided with an initial type:
+Yet again, Santa has made a request to change the children filtering code. This time he just sent an email to the entire engineering team (which is absolutely _not_ the process, but since Santa is sometimes a bit difficult to communicate w, no one has yet had the courage to tell him). Here's the contents of the email.
 
-```typescript
-type RemoveMisbehavingPeople<T> = unknown;
+```http
+POST /sendmail HTTP/1.1
+Host: mail.hohoholdings.com
+Content-Type: text/plain; charset=utf-8
+Content-Length: [Content Length]
+
+From: kris.kringle@hohoholdings.com
+To: engineering@hohoholdings.com
+Subject: Code Changes Needed
+
+Hello beloved team!
+
+Looks like we need some changes to the code again!
+
+1. there are sometimes naughty kids in the same list
+1. turns out I don't actually need to see the nice children in the list, after all
+1. my golf game ran late this morning.. so since the other two changes were quick to implement, I'm sure this will be just as fast!
+
+- Kris Kringle
+  "at Santa's workshop, we value loyalty over all else"
 ```
 
-Your task is to create a TypeScript type `RemoveMisbehavingPeople<T>` that filters out entries from an object where the property key is prepended with `naughty_`. 
+Wow. What a pointless email. For once, calling a meeting would have been better.
 
-Good luck!
+Good thing we got some experience reading the tests because this email may as well have said "do work. thanks." (lol).
+
+Off to the tests to see how this is actually supposed to work!
